+++ src/Pages/landing/LandingContainer.tsx
@@ -1,0 +1,10 @@
+import React, { useState, useEffect } from "react";
+import LandingPresentation from "./LandingPresentation";
+
+const LandingContainer: React.FC = () => {
+  return (
+    <LandingPresentation/>
+  );
+};
+
+export default LandingContainer;
