+++ src/Pages/landing/index.ts
@@ -1,0 +1,3 @@
+import LandingContainer from "./LandingContainer";
+
+export { LandingContainer };
