--- src/Pages/lending/index.ts
@@ -1,3 +1,0 @@
-import LendingContainer from "./LendingContainer";
-
-export { LendingContainer };
