+++ src/Routes/routes.tsx
@@ -1,5 +1,5 @@
 import { RouteObject } from "react-router-dom";
-import { LendingContainer } from "../Pages/lending";
+import { LandingContainer } from "../Pages/landing";
 import { ProjectContainer } from "../Pages/project";
 import { CommunityContainer } from "../Pages/community";
 import { PublicContainer } from "../Pages/public";
@@ -7,16 +7,11 @@ import { PublicContainer } from "../Pages/public";
 export const routes: RouteObject[] = [
   {
     path: "/",
-    element: <LendingContainer />,
+    element: <LandingContainer />,
   },
   {
-    path: "party",
-    children: [
-      {
-        path: ":partyId",
-        element: <ProjectContainer />,
-      },
-    ],
+    path: "project",
+    element: <ProjectContainer />,
   },
   {
     path: "party",
@@ -39,6 +34,6 @@ export const routes: RouteObject[] = [
   },
   {
     path: "*",
-    element: <LendingContainer />,
+    element: <LandingContainer />,
   },
 ];
